[1865] Refactor Cli files to make MCB folder cleaner
