+++ bin/mcb
@@ -24,9 +24,9 @@ require 'mcb/config'
 require 'mcb/course_show'
 require 'mcb/courses_editor'
 require 'mcb/provider_editor'
-require 'mcb/base_cli'
-require 'mcb/courses_editor_cli'
-require 'mcb/provider_cli'
+require 'mcb/cli/base_cli'
+require 'mcb/cli/course_cli'
+require 'mcb/cli/provider_cli'
 require 'mcb/grant_access_wizard'
 require 'mcb/revoke_access_wizard'
 require 'mcb/render'
